Do not try to be backwards-compatible
Always include both resourceName and resourceNamespace in selector.
Previously it was being omitted for backwards-compatibility, but that
creates problems when there are Terraform resources in different
namespaces and also in the same namespace as the controller, both having
the same name.

Just assume that any old resources would have been cleaned up before
upgrading, and if not, force the user to manually delete them.